Meriem (#36)
* create the loginPage

* create forgot password page

* login/forgot pass

* Err solved

* //

* home page added

* login link

* admin pannel init

* sign up ui

* sign-up final version (UI)

* simple text change

* Made updates to backend and package files

* //

* link sign up (demo)

* update

* inscription without files

* /

* last version sign up

* loading behaviour

* validation

* update migrations , contoller and gitignore

* re-inscription

* gestion des absences

* gestion des absences et notifications ajoutees

---------

+++ backend/migrations/20250228214050_add_new_columns.js
@@ -6,7 +6,6 @@ exports.up = async function (knex) {
   await knex.schema.alterTable("users", (table) => {
     table.string("user_role", 255);
   });
-
 };
 
 exports.down = async function (knex) {
@@ -14,7 +13,10 @@ exports.down = async function (knex) {
     table.dropColumn("matricule");
   });
 
-  await knex.schema.alterTable("users", (table) => {
-    table.dropColumn("role");
-  });
+  const hasValidatedColumn = await knex.schema.hasColumn("users", "role");
+  if (hasValidatedColumn) {
+    await knex.schema.alterTable("users", (table) => {
+      table.dropColumn("role");
+    });
+  }
 };
